feat: refactor the firebase.js, break it down to several files
